chore: renamed components/ui dir to components/base-ui
