+++ src/components/custom-sidebar/settings.tsx
@@ -1,13 +1,13 @@
 'use client';
 
 import {ChevronRight, type LucideIcon} from 'lucide-react';
-import {Collapsible, CollapsibleTrigger} from '../ui/collapsible.js';
+import {Collapsible, CollapsibleTrigger} from '../base-ui/collapsible.js';
 import {
 	SidebarGroup,
 	SidebarMenu,
 	SidebarMenuButton,
 	SidebarMenuItem,
-} from '@/components/ui/sidebar.js';
+} from '@/components/base-ui/sidebar.js';
 
 type AllowedPages =
 	| 'Hoare Logic Proof Validator'
